defer bearer creation until javalin startup
